Update Icon New TrayMenu Items Pack some work into funcs

+++ MainPage.cs
@@ -1,21 +1,25 @@
 using System.IO;
 using System.Diagnostics;
 using System.Security.Principal;
+using System.Collections;
 
 namespace SingSelector
 {
     public partial class MainPage : Form
     {
+        
         public MainPage()
         {
             InitializeComponent();
         }
+
 
         bool isSingOn = false;
 
         readonly string CurrentPath = Environment.CurrentDirectory + "\\";
 
         readonly Process SingProc = new();
+
 
         private void MainPage_Load(object sender, EventArgs e)
         {
@@ -35,16 +39,8 @@ namespace SingSelector
                 Environment.Exit(0);
             }
 
-            // 此处的文件带路径
-            string[] LisFiles = Directory.GetFiles(CurrentPath + "Config\\");
-            for (int i = 0; i < LisFiles.Length; i++)
-            {
-                if (LisFiles[i].EndsWith(".json"))
-                {
-                    this.ComboBox_Selector.Items.Add(LisFiles[i].Split('\\')[^1]);
-                }
-            }
-            this.ComboBox_Selector.SelectedItem = ComboBox_Selector.Items[0];
+            // 参见下文
+            Update_Profiles();
 
             SingProc.StartInfo.WorkingDirectory = CurrentPath;
             SingProc.StartInfo.FileName = "sing-box.exe";
@@ -62,45 +58,30 @@ namespace SingSelector
             {
                 isSingOn = false;
                 SingProc.CancelErrorRead();
+
                 this.Button_Switch.Text = "启动";
+                this.TrayMenu_Switch.Text = "启动";
             };
         }
+
 
         #region 按钮点击事件
         private void Button_Switch_Click(object sender, EventArgs e)
         {
-            if (isSingOn)
-            {
-                SingProc.Kill();
-            }
-            else
-            {
-                isSingOn = true;
+            if (isSingOn) SingProc.Kill();
+            else TurnOn_SingBox();
+        }
 
-                this.Button_Switch.Text = "停止";
-                this.RichTextBox_Log.Clear();
-                this.RichTextBox_Log.AppendText("启动: " + this.ComboBox_Selector.SelectedItem + "\n");
-
-                SingProc.StartInfo.Arguments = "run --disable-color -c .\\Config\\" + this.ComboBox_Selector.SelectedItem;
-                SingProc.Start();
-                SingProc.BeginErrorReadLine();
-            }
+        private void TrayMenu_Switch_Click(object sender, EventArgs e)
+        {
+            if (isSingOn) SingProc.Kill();
+            else TurnOn_SingBox();
         }
 
         private void Button_Refresh_Click(object sender, EventArgs e)
         {
-            this.ComboBox_Selector.Items.Clear();
-
-            // 注意这里和初始化时的路径不同
-            string[] LisFiles = Directory.GetFiles(CurrentPath + "Config\\");
-            for (int i = 0; i < LisFiles.Length; i++)
-            {
-                if (LisFiles[i].EndsWith(".json"))
-                {
-                    this.ComboBox_Selector.Items.Add(LisFiles[i].Split('\\')[^1]);
-                }
-            }
-            this.ComboBox_Selector.SelectedItem = ComboBox_Selector.Items[0];
+            if (isSingOn) SingProc.Kill();
+            Update_Profiles();
         }
 
         private void Button_EditProfile_Click(object sender, EventArgs e)
@@ -112,13 +93,22 @@ namespace SingSelector
         }
         #endregion
 
+
         #region 缩小至托盘与展开
         protected override void OnFormClosing(FormClosingEventArgs e)
         {
             if (e.CloseReason == CloseReason.UserClosing)
             {
-                e.Cancel = true;
-                this.Hide();
+                // 非常的银性
+                if (isSingOn)
+                {
+                    e.Cancel = true;
+                    this.Hide();
+                }
+                else
+                {
+                    Environment.Exit(0);
+                }
             }
         }
 
@@ -142,6 +132,7 @@ namespace SingSelector
         }
         #endregion
 
+
         // 文本框自动滚动
         private void RichTextBox_Log_TextChanged(object sender, EventArgs e)
         {
@@ -155,5 +146,101 @@ namespace SingSelector
             if (isSingOn) SingProc.Kill();
         }
 
+
+        #region 复用的一些函数
+
+        // 切换启动&停止
+        private void TurnOn_SingBox()
+        {
+            isSingOn = true;
+
+            // 所有需要变化文字的按钮
+            this.Button_Switch.Text = "停止";
+            this.TrayMenu_Switch.Text = "停止";
+
+            this.RichTextBox_Log.Clear();
+            this.RichTextBox_Log.AppendText("启动: " + this.ComboBox_Selector.SelectedItem + "\n");
+
+            SingProc.StartInfo.Arguments = "run --disable-color -c .\\Config\\" + this.ComboBox_Selector.SelectedItem + ".json";
+            SingProc.Start();
+            SingProc.BeginErrorReadLine();
+        }
+
+        // 刷新配置文件列表
+        private void Update_Profiles()
+        {
+            List<string> profiles = [];
+            this.ComboBox_Selector.Items.Clear();
+            this.TrayMenu_ChangeProfile.DropDownItems.Clear();
+
+            // LisFiles中的文件string带路径
+            string[] LisFiles = Directory.GetFiles(CurrentPath + "Config\\");
+            for (int i = 0; i < LisFiles.Length; i++)
+            {
+                if (LisFiles[i].EndsWith(".json"))
+                {
+                    // 去除路径和后缀
+                    profiles.Add(LisFiles[i].Split('\\')[^1].Split('.')[0]);
+                }
+            }
+
+            // 这里需要检测至少有一个配置文件
+            if (profiles.Count == 0)
+            {
+                MessageBox.Show("未检测到配置文件，请先添加配置文件");
+
+                this.Button_EditProfile.Enabled = false;
+                this.Button_Switch.Enabled = false;
+                this.TrayMenu_ChangeProfile.Enabled = false;
+                this.TrayMenu_Switch.Enabled = false;
+            }
+            else
+            {
+                this.Button_EditProfile.Enabled = true;
+                this.Button_Switch.Enabled = true;
+                this.TrayMenu_ChangeProfile.Enabled = true;
+                this.TrayMenu_Switch.Enabled = true;
+
+                foreach (string item in profiles)
+                {
+                    this.ComboBox_Selector.Items.Add(item);
+                    this.TrayMenu_ChangeProfile.DropDownItems.Add(item);
+                }
+                this.ComboBox_Selector.SelectedItem = ComboBox_Selector.Items[0];
+            }
+        }
+
+        // 从TrayMenu中切换配置时，需要一并修改ComboBox的SelectedItem
+        private void Update_TrayMenu_Selection(object sender, ToolStripItemClickedEventArgs e)
+        {
+            // 为什么这里ClickedItem可能为null啊
+            if (e.ClickedItem != null) this.ComboBox_Selector.SelectedItem = e.ClickedItem.Text;
+
+            // 如果已经启动就重启
+            if (isSingOn)
+            {
+                // 小坑，需要先等待Exited的事件执行完再启动
+                SingProc.Kill();
+                Task.Run(() => {
+                    Thread.Sleep(500);
+                    TurnOn_SingBox();
+                });
+            }
+        }
+
+        // ComboBox切换配置时无需考虑其它
+        private void Update_ComboBox_Selection(object sender, EventArgs e)
+        {
+            if (isSingOn)
+            {
+                SingProc.Kill();
+                Task.Run(() => {
+                    Thread.Sleep(500);
+                    TurnOn_SingBox();
+                });
+            }
+        }
+
+        #endregion
     }
 }
